requirements.txt
Flask==3.0.0
playwright==1.40.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
psutil==5.9.6
bcrypt==4.1.2
